Refactor integrated storage management
